test(identity): dual-fixture IPrincipalStore contract suite
Refactor in-memory store tests into shared abstract cases so the EF
backend is exercised with identical semantics (Testcontainers CI
fail-closed) plus model-only mapping coverage without Docker.
