+++ tests/container-apps/web/web-infrastructure-tests/global-usings.cs
@@ -6,3 +6,4 @@ global using TimeWarp.Architecture.Aggregates.Profiles;
 global using TimeWarp.Architecture.Features.Profiles.Infrastructure;
 global using TimeWarp.Architecture.Persistence;
 global using TimeWarp.Fixie;
+global using TimeWarp.Identity;
